`kernel` Impl exceptions on `aarch64`, many other misc changes

+++ lib/adt/array.h
@@ -43,8 +43,8 @@ void *array_at(struct array array, uint32_t index);
 void *array_front(struct array array);
 void *array_back(struct array array);
 
-uint64_t array_item_count(struct array array);
-uint64_t array_free_count(struct array array);
+uint32_t array_item_count(struct array array);
+uint32_t array_free_count(struct array array);
 
 void *array_take_data(struct array *array);
 void array_take_item(struct array *array, uint32_t index, void *item);
